code/requirements.txt
descartes==1.1.0
matplotlib==3.7.5
networkx==3.1
numpy==1.24.4
opencv_python==4.9.0.80
Pillow==10.3.0
pyemd==1.0.0
pypng==0.20220715.0
scikit_learn==1.3.2
scipy==1.13.1
Shapely==2.0.4
skan==0.11.1
tabulate==0.9.0
scikit-image>=0.21.0